Address hands-on test report: post-message honesty, foreground guard, formatting
Follow-up to Nikola's empirical test report on the input verbs:

- #2/#3: send-keys via post-message warns when literal text is sent, since
  WinUI 3 / XAML drops posted WM_CHAR (typed text silently no-ops). Docs
  corrected to recommend --via send-input for typed text on WinUI/WPF.
- Hardening #1: extend the foreground-ownership guard (was send-keys only)
  to drag and scroll --wheel via a shared ForegroundGuard helper, so an
  OS-wide gesture can't land on the wrong window when SetForegroundWindow
  silently fails. Returns foreground_not_target.
- #6: the "no scroll mode specified" error now uses the same ❌ symbol as
  every other ui error (was raw [ERROR] format).

System-reserved-combo blocklist (report hardening #2) intentionally left
for discussion — it changes the tool contract and partly conflicts with
legitimate automation (testing win+l etc.).
